% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_choice.R
\name{make_mc}
\alias{make_mc}
\title{Make a multiple choice question set}
\usage{
make_mc(
  survey_codebook,
  block_name = NULL,
  response_sep = ";",
  break_after_questions = 0
)
}
\arguments{
\item{survey_codebook}{survey_codebook A dataframe with the question in required format}

\item{block_name}{(optional) Name for a block of questions}

\item{response_sep}{The character for separating response options. Example is semicolon ";". Usage: Yes;No}

\item{break_after_questions}{Add a pagebreak after this many questions}
}
\value{
text for import
}
\description{
Make a multiple choice question set
}
\examples{
#make_mc
}
