% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtricsMaker.R
\docType{package}
\name{qualtricsMaker}
\alias{qualtricsMaker}
\title{Create qualtrics surveys}
\description{
A set of functions to help create Qualtrics Surveys for import.
}
\details{
\tabular{ll}{
Package: \tab qualtricsMaker\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2021-07-11\cr
License: \tab MIT\cr
}
}
\author{
\tabular{ll}{
Author: \tab David J. Stanley \email{dstanley@uoguelph.ca}\cr
Maintainer: \tab David J. Stanley \email{dstanley@uoguelph.ca}
}
}
